Allow compiling and linking to be done on separate platforms
Because of OS cross compile limitations in mojo (see previous commits),
compile actions must be pinned to the target OS. Because the mojo
toolchain was set on the top level rule, all resolved toolchains were
pinned to the target OS, which meant linking could not be cross
compiled. This removes that restriction using exec_groups to free up
macOS CPU time when cross compiling.