Set macOS min OS when passed
Mojo (really LLVM) defaults to the current major OS version if no min OS
version is passed with the triple. In this case if the linker has a
minimum OS version flag it warns (and potentially produces invalid
binaries). This makes sure we're using the correct version for both
cases.